refactor(xapiri): extract renderHelpTab → tab_help.go
Mechanical move per ADR 0014. No behaviour change.
